refactor(ui): rename evalCache.ts to evalOnly.ts
The module is the eval-only read plane (export dump → swagger Flag),
not a general EvalCache client. Name it next to eval.ts / crud.ts and
evalOnlyMode.
